refactor(react-jsx): inline shared rule helpers into call sites
